deduplicate sweep logic
