+++ modules/wallet/scan.go
@@ -1,15 +1,19 @@
 package wallet
 
 import (
+	"errors"
 	"fmt"
 
 	"github.com/NebulousLabs/Sia/build"
 	"github.com/NebulousLabs/Sia/crypto"
 	"github.com/NebulousLabs/Sia/modules"
 	"github.com/NebulousLabs/Sia/types"
+	"github.com/NebulousLabs/bolt"
 )
 
 const scanMultiplier = 4 // how many more keys to generate after each scan iteration
+
+const outputSize = 350 // approx. size in bytes of an output and accompanying signature
 
 // numInitialKeys is the number of keys generated by the seedScanner before
 // scanning the blockchain for the first time.
@@ -242,4 +246,130 @@ func new033xScanner(savedKeys []savedKey033x) *v033xScanner {
 		siacoinOutputs: make(map[types.SiacoinOutputID]scannedOutput),
 		siafundOutputs: make(map[types.SiafundOutputID]scannedOutput),
 	}
+}
+
+// sweepOutputs sweeps the supplied siacoin and siafund outputs into the wallet.
+func (w *Wallet) sweepOutputs(scos, sfos []scannedOutput) (coins, funds types.Currency, err error) {
+	// get an address to spend into
+	var uc types.UnlockConditions
+	err = w.db.Update(func(tx *bolt.Tx) error {
+		var err error
+		uc, err = w.nextPrimarySeedAddress(tx)
+		return err
+	})
+	if err != nil {
+		return
+	}
+
+	// construct a transaction that spends the outputs
+	// TODO: this may result in transactions that are too large.
+	tb := w.StartTransaction()
+	defer func() {
+		if err != nil {
+			tb.Drop()
+		}
+	}()
+	var sweptCoins, sweptFunds types.Currency // total values of swept outputs
+	for _, output := range scos {
+		// construct a siacoin input that spends the output
+		tb.AddSiacoinInput(types.SiacoinInput{
+			ParentID:         types.SiacoinOutputID(output.id),
+			UnlockConditions: output.spendableKey.UnlockConditions,
+		})
+		// add a signature for the input
+		sweptCoins = sweptCoins.Add(output.value)
+	}
+	for _, output := range sfos {
+		// construct a siafund input that spends the output
+		tb.AddSiafundInput(types.SiafundInput{
+			ParentID:         types.SiafundOutputID(output.id),
+			UnlockConditions: output.spendableKey.UnlockConditions,
+		})
+		// add a signature for the input
+		sweptFunds = sweptFunds.Add(output.value)
+	}
+
+	// estimate the transaction size and fee. NOTE: this equation doesn't
+	// account for other fields in the transaction, but since we are
+	// multiplying by maxFee, lowballing is ok
+	estTxnSize := (len(scos) + len(sfos)) * outputSize
+	_, maxFee := w.tpool.FeeEstimation()
+	estFee := maxFee.Mul64(uint64(estTxnSize))
+	tb.AddMinerFee(estFee)
+
+	// calculate total siacoin payout
+	if sweptCoins.Cmp(estFee) > 0 {
+		coins = sweptCoins.Sub(estFee)
+	}
+	funds = sweptFunds
+
+	switch {
+	case coins.IsZero() && funds.IsZero():
+		// if we aren't sweeping any coins or funds, then just return an
+		// error; no reason to proceed
+		return types.Currency{}, types.Currency{}, errors.New("transaction fee exceeds value of swept outputs")
+
+	case !coins.IsZero() && funds.IsZero():
+		// if we're sweeping coins but not funds, add a siacoin output for
+		// them
+		tb.AddSiacoinOutput(types.SiacoinOutput{
+			Value:      coins,
+			UnlockHash: uc.UnlockHash(),
+		})
+
+	case coins.IsZero() && !funds.IsZero():
+		// if we're sweeping funds but not coins, add a siafund output for
+		// them. This is tricky because we still need to pay for the
+		// transaction fee, but we can't simply subtract the fee from the
+		// output value like we can with swept coins. Instead, we need to fund
+		// the fee using the existing wallet balance.
+		tb.AddSiafundOutput(types.SiafundOutput{
+			Value:      funds,
+			UnlockHash: uc.UnlockHash(),
+		})
+		err = tb.FundSiacoins(estFee)
+		if err != nil {
+			return types.Currency{}, types.Currency{}, errors.New("couldn't pay transaction fee on swept funds: " + err.Error())
+		}
+
+	case !coins.IsZero() && !funds.IsZero():
+		// if we're sweeping both coins and funds, add a siacoin output and a
+		// siafund output
+		tb.AddSiacoinOutput(types.SiacoinOutput{
+			Value:      coins,
+			UnlockHash: uc.UnlockHash(),
+		})
+		tb.AddSiafundOutput(types.SiafundOutput{
+			Value:      funds,
+			UnlockHash: uc.UnlockHash(),
+		})
+	}
+
+	// add signatures for all coins and funds (manually, since tb doesn't have
+	// access to the signing keys)
+	txn, parents := tb.View()
+	for _, output := range scos {
+		sk := output.spendableKey
+		addSignatures(&txn, types.FullCoveredFields, sk.UnlockConditions, crypto.Hash(output.id), sk)
+	}
+	for _, output := range sfos {
+		sk := output.spendableKey
+		addSignatures(&txn, types.FullCoveredFields, sk.UnlockConditions, crypto.Hash(output.id), sk)
+	}
+	// Usually, all the inputs will come from swept outputs. However, there is
+	// an edge case in which inputs will be added from the wallet. To cover
+	// this case, we iterate through the SiacoinInputs and add a signature for
+	// any input that belongs to the wallet.
+	w.mu.RLock()
+	for _, input := range txn.SiacoinInputs {
+		if key, ok := w.keys[input.UnlockConditions.UnlockHash()]; ok {
+			addSignatures(&txn, types.FullCoveredFields, input.UnlockConditions, crypto.Hash(input.ParentID), key)
+		}
+	}
+	w.mu.RUnlock()
+
+	// submit the transaction
+	txnSet := append(parents, txn)
+	err = w.tpool.AcceptTransactionSet(txnSet)
+	return
 }
